pyproject.toml
[tool.poetry]
name = "hermes"
version = "0.0.0"
description = "Backend of a trading platform between sports hospitality vendors"
authors = ["Yifei Yu"]
readme = "README.md"
packages = [
    { include = "hermes", from = "src" }
]
include = [
    { path = "src/hermes/conf_default", format = "sdist" },
    { path = "notebooks", format = "sdist" }
]

[tool.poetry.dependencies]
python = "~3.11"
pydantic = "^2.8.2"
pandas = "^2.2.2"
fastapi = "^0.111.1"
email-validator = "^2.2.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = {extras = ["mypy"], version = "^2.0.31"}
streamlit = "^1.37.0"
bcrypt = "^4.2.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pytest = "^8.2.2"
mypy = "^1.10.1"
flake8 = "^7.1.0"
coverage = {extras = ["toml"], version = "^7.6.0"}
black = "^24.4.2"
yamllint = "^1.35.1"
semgrep = "^1.79.0"
ipykernel = "^6.29.5"
nbformat = "^5.10.4"
ruff = "^0.5.2"
shellcheck-py = "^0.10.0.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
force-exclude = '''
.*/setup\.py$
'''

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests"]
extend_skip = ["setup.py"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--junitxml=test_report.xml"
testpaths = "tests"

[tool.mypy]
files = "src"
mypy_path = "src"
exclude = [
    ".venv"
]
plugins = "sqlalchemy.ext.mypy.plugin"
namespace_packages = true
explicit_package_bases = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
warn_return_any = false

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",
    '''if self\.debug''',

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    '''@(abc\.)?abstractmethod''',
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
